requirements.txt
django
channels["daphne"]
djangorestframework
Pillow